requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
python-multipart>=0.0.6
Pillow>=9.5.0
torch>=2.0.0
git+https://github.com/openai/CLIP.git
google-cloud-storage>=2.8.0  # Include GCS client library
chromadb>=0.4.6
pydantic>=1.10.7
numpy>=1.24.2
python-dotenv>=1.0.0
google-genai
pydantic-settings
psycopg2-binary